refactor: Remove useless ComposeEditor view
